Fixed checking for do not disturb time
When DND spans midnight, I need to move the end time to tomorrow if the end time is before the start time and now is before midnight, and move start time to yesterday when end time is before start time and now is after midnight. Then a straight is now is between start time and end time check will work.